# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NYX API
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /v1/flows:
    get:
      summary: List flows
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - in: query
          name: after
          schema:
            type: string
      responses:
        "200":
          description: Paginated flow list
    post:
      summary: Create a flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, target, objective]
              properties:
                name:
                  type: string
                target:
                  type: string
                  format: uri
                objective:
                  type: string
      responses:
        "201":
          description: Flow created
        "400":
          description: Validation failed
  /v1/flows/{flowId}:
    get:
      summary: Get flow detail
      parameters:
        - in: path
          name: flowId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Flow detail
...